deleting of annotations
The DELETE annotations/:id endpoint deletes annotations via the update handler.
Because there is actually no way to send the rev number of a doc as header or query param (via the annotateit.org widget / method "destroy") - the update handler simply add the param `_deleted: true`.

That was the last needed Operation to support the full CRUD circle.